+++ CHANGELOG.md
@@ -73,6 +73,7 @@ Please view this file on the master branch, on stable branches it's out of date.
 - Autoselect default project to open access service order and go directly to summary page (@mkasztelnik)
 - Cannot order service when there is not offers (@mkasztelnik)
 - Related services (@mkasztelnik)
+- Service activate message (@mkasztelnik)
 
 ### Changed
 - Upgrade Sprockets gem to avoid CVE-2018-3760 vulnerability (@mkasztelnik)